Реорганизация actions
